Refactored name to CreateUserConversationService
